Fix: Move integration tests
