+++ tests/Integration/Http/Controller/Admin/TalksControllerTest.php
@@ -16,7 +16,7 @@ namespace OpenCFP\Test\Integration\Http\Controller\Admin;
 use OpenCFP\Domain\Model\Talk;
 use OpenCFP\Domain\Model\TalkMeta;
 use OpenCFP\Test\Helper\RefreshDatabase;
-use OpenCFP\Test\WebTestCase;
+use OpenCFP\Test\Integration\WebTestCase;
 
 /**
  * @coversNothing
