+++ tests/Integration/Http/Controller/DashboardControllerTest.php
@@ -17,7 +17,7 @@ use Mockery as m;
 use OpenCFP\Domain\Services\Authentication;
 use OpenCFP\Domain\Speaker\SpeakerProfile;
 use OpenCFP\Infrastructure\Auth\UserInterface;
-use OpenCFP\Test\WebTestCase;
+use OpenCFP\Test\Integration\WebTestCase;
 
 /**
  * @group db
