+++ tests/Integration/Http/Controller/ProfileControllerTest.php
@@ -15,7 +15,7 @@ namespace OpenCFP\Test\Integration\Http\Controller;
 
 use OpenCFP\Domain\Model\User;
 use OpenCFP\Test\Helper\RefreshDatabase;
-use OpenCFP\Test\WebTestCase;
+use OpenCFP\Test\Integration\WebTestCase;
 
 /**
  * @group db
