+++ tests/Integration/Infrastructure/Event/Fixtures/global_with_admin.txt
