+++ tests/Integration/Infrastructure/Event/Fixtures/global_with_anonymous_user.txt
