+++ tests/Integration/Infrastructure/Templating/Fixtures/functions.txt
